Add a simple kafka + zookeeper configuration
The patch adds a simple docker compose configuration for kafka +
zookeeper. In addition to them, the are two more services: manager and
redpanda. The redpanda service is a GUI to manage a kafka cluster and
the manager container is a jupyter notebook that comes with a bash
kernel, so it gets easy to prepare the input commands for demo. The
manager container communicate with kafka through a ssh session so they
share ssh keys.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>